% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data}
\format{
## `fire`
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{x 1000 dollars in damage}
  ...
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
