% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinomial.R
\name{mybin}
\alias{mybin}
\title{Binomial Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{sample size to take}

\item{p}{probabilty}
}
\value{
a table of the number of success and a barplot of the number of successes
}
\description{
Binomial Function
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
