% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations in which the function will take samples}

\item{x}{the sample in which the function is sampling from}

\item{fun}{the function or the statistic the function is sampling for}

\item{alpha}{1 - the confidence interval percentage}

\item{cx}{the font size of the text}

\item{...}{ylab,xlab, etc..}
}
\value{
a list containing pertinent info the bootstrap function has measured
}
\description{
Bootstrap Function
}
\examples{
set.seed(39)
sam=rnorm(25,mean=25,sd=10)
myboot2(iter=1000,x=sam,fun="mean",alpha=0.05)
}
