% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{lower bound limit}

\item{b}{upper bound limit}
}
\value{
a histogram displaying the runif sample made with the given params
}
\description{
mycltu
}
\examples{
mycltu(30, 10000, 0, 10)
}
