% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{mycurve}
\alias{mycurve}
\title{Title}
\usage{
mycurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{the probability}
}
\value{
graphs the curve givem the params
}
\description{
Title
}
\examples{
mycurve(5, 2, 7)
}
