% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{My max likelihood norm function}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample}

\item{mu}{mus to test}

\item{sig}{sigmas to test}

\item{...}{}
}
\value{
a graphical depiction of the estimate of the maximum likelihood value
}
\description{
My max likelihood norm function
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
