% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysq.R
\name{mysq}
\alias{mysq}
\title{MySquare Function}
\usage{
mysq(x)
}
\arguments{
\item{x}{a quantitative vector}
}
\value{
A quantitative vector with square components
}
\description{
MySquare Function
}
\examples{
mysq(1:10)
}
