% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{N Tickets Function for maximizing ticket sales}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats available}

\item{gamma}{desired probability of overbooking}

\item{p}{probability that a passenger who purchased a ticket will show up}
}
\value{
A discrete plot and a continuous plot, a list printing the discrete and continuous optimal values, the probability, and the gamma
}
\description{
N Tickets Function for maximizing ticket sales
}
\examples{
ntickets(400, 0.02, 0.95)

}
