% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter Histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{X var as a vector}

\item{y}{Y var as a vector}

\item{xlab}{title of the x axis}

\item{ylab}{title of the y axis}
}
\value{
A scatter plot histogram
}
\description{
Scatter Histogram
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
